Module pallet_chainlink_feed::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_default_parts_6 as tt_default_parts;

Structs

Feed

Proxy used for interaction with a feed. should_sync flag determines whether the config is put into storage on drop.

FeedBuilder

Builder with all the parameters to call the create_feed

FeedConfig

The configuration for an oracle feed.

GenesisConfig

Can be used to configure the genesis state of this pallet.

OracleMeta

Meta data tracking withdrawable rewards and admin for an oracle.

OracleStatus

Meta data tracking the oracle status for a feed.

Pallet

The pallet implementing the on-chain logic.

Requester

Used to store round requester permissions for accounts.

Round

Round data relevant to consumers. Will only be constructed once minimum amount of submissions have been provided.

RoundData

Round data as served by the FeedInterface.

RoundDetails

Round data relevant to oracles.

_GeneratedPrefixForStorageDetails

_GeneratedPrefixForStorageFeedCounter

_GeneratedPrefixForStorageFeedCreators

_GeneratedPrefixForStorageFeeds

_GeneratedPrefixForStorageOracleStatuses

_GeneratedPrefixForStorageOracles

_GeneratedPrefixForStoragePalletAdmin

_GeneratedPrefixForStoragePendingPalletAdmin

_GeneratedPrefixForStorageRequesters

_GeneratedPrefixForStorageRounds

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

RoundConversionError

Possible error when converting from Round to RoundData.

Traits

Config

Configuration trait of this pallet.

FeedInterface

Trait for read-only access to a feed.

FeedOracle

Trait for interacting with the feeds in the pallet.

MutableFeedInterface

Trait for read-write access to a feed.

WeightInfo

Trait for the chainlink pallet extrinsic weights.

Type Definitions

BalanceOf

Details

Operator-facing round data.

FeedBuilderOf

FeedConfigOf

FeedCounter

A running counter used internally to determine the next feed id.

FeedCreators

Accounts allowed to create feeds.

Feeds

A running counter used internally to determine the next feed id.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

OracleMetaOf

OracleStatusOf

OracleStatuses

Feed local oracle status data.

Oracles

Global oracle meta data including admin and withdrawable funds.

PalletAdmin

The account controlling the funds for this pallet.

PendingPalletAdmin

The account to set as future pallet admin.

Requesters

Per-feed permissioning for starting new rounds.

RoundDataOf

RoundDetailsOf

RoundId

RoundOf

Rounds

User-facing round data.

SubmissionBounds

Minimum and Maximum number of submissions allowed per round.